Split out rectangle and outline stuff
