+++ PyPDF2/generic/_data_structures.py
@@ -29,10 +29,8 @@
 __author__ = "Mathieu Fenniak"
 __author_email__ = "biziqe@mathieu.fenniak.net"
 
-import decimal
 import logging
 import re
-from enum import IntFlag
 from io import BytesIO
 from typing import Any, Dict, Iterable, List, Optional, Tuple, Union, cast
 
@@ -40,7 +38,6 @@ from .._utils import (
     WHITESPACES,
     StreamType,
     b_,
-    deprecate_no_replacement,
     deprecate_with_replacement,
     hex_str,
     logger_warning,
@@ -53,6 +50,7 @@ from ..constants import (
     FieldDictionaryAttributes,
 )
 from ..constants import FilterTypes as FT
+from ..constants import OutlineFontFlag
 from ..constants import StreamAttributes as SA
 from ..constants import TypArguments as TA
 from ..constants import TypFitArguments as TF
@@ -837,249 +835,6 @@ def read_object(
             return NumberObject.read_from_stream(stream)
 
 
-class RectangleObject(ArrayObject):
-    """
-    This class is used to represent *page boxes* in PyPDF2. These boxes include:
-        * :attr:`artbox <PyPDF2._page.PageObject.artbox>`
-        * :attr:`bleedbox <PyPDF2._page.PageObject.bleedbox>`
-        * :attr:`cropbox <PyPDF2._page.PageObject.cropbox>`
-        * :attr:`mediabox <PyPDF2._page.PageObject.mediabox>`
-        * :attr:`trimbox <PyPDF2._page.PageObject.trimbox>`
-    """
-
-    def __init__(
-        self, arr: Union["RectangleObject", Tuple[float, float, float, float]]
-    ) -> None:
-        # must have four points
-        assert len(arr) == 4
-        # automatically convert arr[x] into NumberObject(arr[x]) if necessary
-        ArrayObject.__init__(self, [self._ensure_is_number(x) for x in arr])  # type: ignore
-
-    def _ensure_is_number(self, value: Any) -> Union[FloatObject, NumberObject]:
-        if not isinstance(value, (NumberObject, FloatObject)):
-            value = FloatObject(value)
-        return value
-
-    def scale(self, sx: float, sy: float) -> "RectangleObject":
-        return RectangleObject(
-            (
-                float(self.left) * sx,
-                float(self.bottom) * sy,
-                float(self.right) * sx,
-                float(self.top) * sy,
-            )
-        )
-
-    def ensureIsNumber(
-        self, value: Any
-    ) -> Union[FloatObject, NumberObject]:  # pragma: no cover
-        deprecate_no_replacement("ensureIsNumber")
-        return self._ensure_is_number(value)
-
-    def __repr__(self) -> str:
-        return f"RectangleObject({repr(list(self))})"
-
-    @property
-    def left(self) -> FloatObject:
-        return self[0]
-
-    @property
-    def bottom(self) -> FloatObject:
-        return self[1]
-
-    @property
-    def right(self) -> FloatObject:
-        return self[2]
-
-    @property
-    def top(self) -> FloatObject:
-        return self[3]
-
-    def getLowerLeft_x(self) -> FloatObject:  # pragma: no cover
-        deprecate_with_replacement("getLowerLeft_x", "left")
-        return self.left
-
-    def getLowerLeft_y(self) -> FloatObject:  # pragma: no cover
-        deprecate_with_replacement("getLowerLeft_y", "bottom")
-        return self.bottom
-
-    def getUpperRight_x(self) -> FloatObject:  # pragma: no cover
-        deprecate_with_replacement("getUpperRight_x", "right")
-        return self.right
-
-    def getUpperRight_y(self) -> FloatObject:  # pragma: no cover
-        deprecate_with_replacement("getUpperRight_y", "top")
-        return self.top
-
-    def getUpperLeft_x(self) -> FloatObject:  # pragma: no cover
-        deprecate_with_replacement("getUpperLeft_x", "left")
-        return self.left
-
-    def getUpperLeft_y(self) -> FloatObject:  # pragma: no cover
-        deprecate_with_replacement("getUpperLeft_y", "top")
-        return self.top
-
-    def getLowerRight_x(self) -> FloatObject:  # pragma: no cover
-        deprecate_with_replacement("getLowerRight_x", "right")
-        return self.right
-
-    def getLowerRight_y(self) -> FloatObject:  # pragma: no cover
-        deprecate_with_replacement("getLowerRight_y", "bottom")
-        return self.bottom
-
-    @property
-    def lower_left(self) -> Tuple[decimal.Decimal, decimal.Decimal]:
-        """
-        Property to read and modify the lower left coordinate of this box
-        in (x,y) form.
-        """
-        return self.left, self.bottom
-
-    @lower_left.setter
-    def lower_left(self, value: List[Any]) -> None:
-        self[0], self[1] = (self._ensure_is_number(x) for x in value)
-
-    @property
-    def lower_right(self) -> Tuple[decimal.Decimal, decimal.Decimal]:
-        """
-        Property to read and modify the lower right coordinate of this box
-        in (x,y) form.
-        """
-        return self.right, self.bottom
-
-    @lower_right.setter
-    def lower_right(self, value: List[Any]) -> None:
-        self[2], self[1] = (self._ensure_is_number(x) for x in value)
-
-    @property
-    def upper_left(self) -> Tuple[decimal.Decimal, decimal.Decimal]:
-        """
-        Property to read and modify the upper left coordinate of this box
-        in (x,y) form.
-        """
-        return self.left, self.top
-
-    @upper_left.setter
-    def upper_left(self, value: List[Any]) -> None:
-        self[0], self[3] = (self._ensure_is_number(x) for x in value)
-
-    @property
-    def upper_right(self) -> Tuple[decimal.Decimal, decimal.Decimal]:
-        """
-        Property to read and modify the upper right coordinate of this box
-        in (x,y) form.
-        """
-        return self.right, self.top
-
-    @upper_right.setter
-    def upper_right(self, value: List[Any]) -> None:
-        self[2], self[3] = (self._ensure_is_number(x) for x in value)
-
-    def getLowerLeft(
-        self,
-    ) -> Tuple[decimal.Decimal, decimal.Decimal]:  # pragma: no cover
-        deprecate_with_replacement("getLowerLeft", "lower_left")
-        return self.lower_left
-
-    def getLowerRight(
-        self,
-    ) -> Tuple[decimal.Decimal, decimal.Decimal]:  # pragma: no cover
-        deprecate_with_replacement("getLowerRight", "lower_right")
-        return self.lower_right
-
-    def getUpperLeft(
-        self,
-    ) -> Tuple[decimal.Decimal, decimal.Decimal]:  # pragma: no cover
-        deprecate_with_replacement("getUpperLeft", "upper_left")
-        return self.upper_left
-
-    def getUpperRight(
-        self,
-    ) -> Tuple[decimal.Decimal, decimal.Decimal]:  # pragma: no cover
-        deprecate_with_replacement("getUpperRight", "upper_right")
-        return self.upper_right
-
-    def setLowerLeft(self, value: Tuple[float, float]) -> None:  # pragma: no cover
-        deprecate_with_replacement("setLowerLeft", "lower_left")
-        self.lower_left = value  # type: ignore
-
-    def setLowerRight(self, value: Tuple[float, float]) -> None:  # pragma: no cover
-        deprecate_with_replacement("setLowerRight", "lower_right")
-        self[2], self[1] = (self._ensure_is_number(x) for x in value)
-
-    def setUpperLeft(self, value: Tuple[float, float]) -> None:  # pragma: no cover
-        deprecate_with_replacement("setUpperLeft", "upper_left")
-        self[0], self[3] = (self._ensure_is_number(x) for x in value)
-
-    def setUpperRight(self, value: Tuple[float, float]) -> None:  # pragma: no cover
-        deprecate_with_replacement("setUpperRight", "upper_right")
-        self[2], self[3] = (self._ensure_is_number(x) for x in value)
-
-    @property
-    def width(self) -> decimal.Decimal:
-        return self.right - self.left
-
-    def getWidth(self) -> decimal.Decimal:  # pragma: no cover
-        deprecate_with_replacement("getWidth", "width")
-        return self.width
-
-    @property
-    def height(self) -> decimal.Decimal:
-        return self.top - self.bottom
-
-    def getHeight(self) -> decimal.Decimal:  # pragma: no cover
-        deprecate_with_replacement("getHeight", "height")
-        return self.height
-
-    @property
-    def lowerLeft(self) -> Tuple[decimal.Decimal, decimal.Decimal]:  # pragma: no cover
-        deprecate_with_replacement("lowerLeft", "lower_left")
-        return self.lower_left
-
-    @lowerLeft.setter
-    def lowerLeft(
-        self, value: Tuple[decimal.Decimal, decimal.Decimal]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("lowerLeft", "lower_left")
-        self.lower_left = value
-
-    @property
-    def lowerRight(self) -> Tuple[decimal.Decimal, decimal.Decimal]:  # pragma: no cover
-        deprecate_with_replacement("lowerRight", "lower_right")
-        return self.lower_right
-
-    @lowerRight.setter
-    def lowerRight(
-        self, value: Tuple[decimal.Decimal, decimal.Decimal]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("lowerRight", "lower_right")
-        self.lower_right = value
-
-    @property
-    def upperLeft(self) -> Tuple[decimal.Decimal, decimal.Decimal]:  # pragma: no cover
-        deprecate_with_replacement("upperLeft", "upper_left")
-        return self.upper_left
-
-    @upperLeft.setter
-    def upperLeft(
-        self, value: Tuple[decimal.Decimal, decimal.Decimal]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("upperLeft", "upper_left")
-        self.upper_left = value
-
-    @property
-    def upperRight(self) -> Tuple[decimal.Decimal, decimal.Decimal]:  # pragma: no cover
-        deprecate_with_replacement("upperRight", "upper_right")
-        return self.upper_right
-
-    @upperRight.setter
-    def upperRight(
-        self, value: Tuple[decimal.Decimal, decimal.Decimal]
-    ) -> None:  # pragma: no cover
-        deprecate_with_replacement("upperRight", "upper_right")
-        self.upper_right = value
-
-
 class Field(TreeObject):
     """
     A class representing a field dictionary.
@@ -1214,15 +969,6 @@ class Field(TreeObject):
         """
         deprecate_with_replacement("additionalActions", "additional_actions")
         return self.additional_actions
-
-
-class OutlineFontFlag(IntFlag):
-    """
-    A class used as an enumerable flag for formatting an outline font
-    """
-
-    italic = 1
-    bold = 2
 
 
 class Destination(TreeObject):
@@ -1399,33 +1145,3 @@ class Destination(TreeObject):
         absolute value = number of visible descendents at all levels
         """
         return self.get("/Count", None)
-
-
-class OutlineItem(Destination):
-    def write_to_stream(
-        self, stream: StreamType, encryption_key: Union[None, str, bytes]
-    ) -> None:
-        stream.write(b"<<\n")
-        for key in [
-            NameObject(x)
-            for x in ["/Title", "/Parent", "/First", "/Last", "/Next", "/Prev"]
-            if x in self
-        ]:
-            key.write_to_stream(stream, encryption_key)
-            stream.write(b" ")
-            value = self.raw_get(key)
-            value.write_to_stream(stream, encryption_key)
-            stream.write(b"\n")
-        key = NameObject("/Dest")
-        key.write_to_stream(stream, encryption_key)
-        stream.write(b" ")
-        value = self.dest_array
-        value.write_to_stream(stream, encryption_key)
-        stream.write(b"\n")
-        stream.write(b">>")
-
-
-class Bookmark(OutlineItem):  # pragma: no cover
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        deprecate_with_replacement("Bookmark", "OutlineItem")
-        super().__init__(*args, **kwargs)
