+++ PyPDF2/types.py
@@ -15,7 +15,8 @@ except ImportError:
     from typing_extensions import TypeAlias
 
 from .generic._base import NameObject, NullObject, NumberObject
-from .generic._data_structures import ArrayObject, Destination, OutlineItem
+from .generic._data_structures import ArrayObject, Destination
+from .generic._outline import OutlineItem
 
 BorderArrayType: TypeAlias = List[Union[NameObject, NumberObject, ArrayObject]]
 OutlineItemType: TypeAlias = Union[OutlineItem, Destination]
